Reset path if enemy gets to last seen player position
When an enemy sees the player it will go towards the player position. If the player goes out of sight, range, etc the enemy will go to the last known position. Once it gets to that position it will stop and not have its path set to the position anymore. this will help when we add a wandering mode so the last known position doesn't interfere